requirements.txt
mlflow==2.4
jupyter
scikit-learn==1.2.2
click
pandas
pyarrow
python-dotenv==1.0.0
google-cloud-storage
prefect
gcsfs
functions-framework
google-cloud-pubsub